+++ internal/provider/resource_sdwan_cisco_system_feature_template_test.go
@@ -69,9 +69,6 @@ func TestAccSdwanCiscoSystemFeatureTemplate(t *testing.T) {
 					resource.TestCheckResourceAttr("sdwan_cisco_system_feature_template.test", "idle_timeout", "100"),
 					resource.TestCheckResourceAttr("sdwan_cisco_system_feature_template.test", "trackers.0.name", "tracker1"),
 					resource.TestCheckResourceAttr("sdwan_cisco_system_feature_template.test", "trackers.0.endpoint_ip", "5.6.7.8"),
-					resource.TestCheckResourceAttr("sdwan_cisco_system_feature_template.test", "trackers.0.transport_endpoint_ip", "5.6.7.8"),
-					resource.TestCheckResourceAttr("sdwan_cisco_system_feature_template.test", "trackers.0.transport_endpoint_protocol", "tcp"),
-					resource.TestCheckResourceAttr("sdwan_cisco_system_feature_template.test", "trackers.0.transport_endpoint_port", "500"),
 					resource.TestCheckResourceAttr("sdwan_cisco_system_feature_template.test", "trackers.0.endpoint_dns_name", "abc.com"),
 					resource.TestCheckResourceAttr("sdwan_cisco_system_feature_template.test", "trackers.0.endpoint_api_url", "https://1.1.1.1"),
 					resource.TestCheckResourceAttr("sdwan_cisco_system_feature_template.test", "trackers.0.boolean", "or"),
@@ -151,9 +148,6 @@ func testAccSdwanCiscoSystemFeatureTemplateConfig_all() string {
 		trackers = [{
 			name = "tracker1"
 			endpoint_ip = "5.6.7.8"
-			transport_endpoint_ip = "5.6.7.8"
-			transport_endpoint_protocol = "tcp"
-			transport_endpoint_port = 500
 			endpoint_dns_name = "abc.com"
 			endpoint_api_url = "https://1.1.1.1"
 			elements = ["abc", "def"]